refactor get_header and submit_block into modules for readability
