+++ config.example.toml
@@ -12,9 +12,6 @@ chain = "Holesky"
 # Docker image to use for the PBS module.
 # OPTIONAL, DEFAULT: ghcr.io/commit-boost/pbs:latest
 docker_image = "ghcr.io/commit-boost/pbs:latest"
-# Whether to enable the PBS module to request signatures from the Signer module (not used in the default PBS image)
-# OPTIONAL, DEFAULT: false
-with_signer = false
 # Host to receive BuilderAPI calls from beacon node
 # OPTIONAL, DEFAULT: 127.0.0.1
 host = "127.0.0.1"
@@ -174,7 +171,7 @@ timeout_get_header_ms = 900
 id = "mux-relay-1"
 url = "http://0xa119589bb33ef52acbb8116832bec2b58fca590fe5c85eac5d3230b44d5bc09fe73ccd21f88eab31d6de16194d17782e@def.xyz"
 
-# Configuration for the Signer Module, only required if any `commit` module is present, or if `pbs.with_signer = true`
+# Configuration for the Signer Module, only required if any `commit` module is present
 # Currently three types of Signer modules are supported (only one can be used at a time):
 #   - Remote: a remote Web3Signer instance
 #   - Dirk: a remote Dirk instance
